# TODO-PLAN/cloudbox-openapi.yaml
openapi: 3.0.3
info:
  title: CloudBox API
  version: 1.0.0
  description: 私有文件管理与分享系统 V1 REST API。
servers:
  - url: http://localhost:8090/api/v1
    description: Local development
...
paths:
...
  /users/me:
    get:
      tags: [Users]
      operationId: getCurrentUser
      summary: 获取当前用户
      responses:
        '200':
          description: 当前用户
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
    patch:
      tags: [Users]
      operationId: updateCurrentUser
      summary: 修改当前用户资料
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      responses:
        '200':
          description: 修改后的用户
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
  /users/me/password:
    put:
      tags: [Users]
      operationId: changePassword
      summary: 修改密码
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '200':
          description: 修改成功，服务端应吊销该用户所有 Refresh Token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ShareToken:
      type: apiKey
      in: header
      name: X-Share-Token
...
  responses:
...
    BadRequest:
      description: 请求参数或操作不合法
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: 未认证或凭据无效
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    Conflict:
      description: 资源状态或名称冲突
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    UpdateProfileRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
          maxLength: 128
    ChangePasswordRequest:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
          format: password
        newPassword:
          type: string
          format: password
          minLength: 8
          maxLength: 72
...
    UserStatus:
      type: string
      enum: [ACTIVE, DISABLED]
    UserRole:
      type: string
      enum: [USER, ADMIN]
...
    User:
      type: object
      required: [id, username, email, role, status, quotaBytes, usedBytes, createdAt]
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        email:
          type: string
          format: email
        role:
          $ref: '#/components/schemas/UserRole'
        status:
          $ref: '#/components/schemas/UserStatus'
        quotaBytes:
          type: integer
          format: int64
        usedBytes:
          type: integer
          format: int64
        createdAt:
          type: string
          format: date-time
...
    ApiResponseVoid:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: OK
        message:
          type: string
          example: success
        data:
          nullable: true
          example: null
    UserResponse:
      type: object
      required: [code, message, data]
      properties:
        code:
          type: string
          example: OK
        message:
          type: string
          example: success
        data:
          $ref: '#/components/schemas/User'
...
    ErrorDetail:
      type: object
      properties:
        field:
          type: string
        message:
          type: string
    ErrorResponse:
      type: object
      required: [code, message, requestId]
      properties:
        code:
          type: string
          example: FILE_NAME_CONFLICT
        message:
          type: string
          example: 目标目录中已存在同名文件
        requestId:
          type: string
        errors:
          type: array
          nullable: true
          items:
            $ref: '#/components/schemas/ErrorDetail'